requirements.txt
fastapi==0.116.1
langchain==0.3.27
langchain_community==0.3.27
langchain_google_vertexai==2.0.28
pydantic==2.11.7
python-dotenv==1.1.1
uvicorn==0.35.0
faiss-cpu==1.11.0.post1
